feat: organize files for modernjs
